+++ src/main/java/org/geowe/service/geometry/JtsSymDifferenceService.java
@@ -65,8 +65,7 @@ public class JtsSymDifferenceService {
 	public Response getDifferenceElements(@NotNull @Valid OperationData operationData,
 			@QueryParam("tolerance") @DefaultValue("-0.00001") double tolerance) {
 		GeoEngineer geoEngineer = new JTSGeoEngineer();
-
-		List<String> elements = geoEngineer.calculateSymDifferenceElements(operationData, tolerance);
+		List<FlatGeometry> elements = geoEngineer.calculateSymDifferenceElements(operationData, tolerance);
 
 		return Response.status(Status.CREATED).entity(elements).build();
 	}