+++ src/test/java/org/geowe/service/geometry/test/DifferenceTest.java
@@ -113,8 +113,9 @@ public class DifferenceTest {
 		Response response = target.request().post(
 				Entity.entity(opData,"application/json;charset=UTF-8"));
 
-		List<String> differenceGeometries = response.readEntity(new GenericType<List<String>>(){});
+		List<FlatGeometry> differenceGeometries = response.readEntity(new GenericType<List<FlatGeometry>>(){});
 		response.close();
+		log.info(differenceGeometries);
 		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
 		Assert.isTrue(differenceGeometries.size() == 10);
 	}
@@ -126,8 +127,9 @@ public class DifferenceTest {
 		Response response = target.queryParam("tolerance", 0.0001).request().post(
 				Entity.entity(opData,"application/json;charset=UTF-8"));
 
-		List<String> differenceGeometries = response.readEntity(new GenericType<List<String>>(){});
+		List<FlatGeometry> differenceGeometries = response.readEntity(new GenericType<List<FlatGeometry>>(){});
 		response.close();
+		log.info(differenceGeometries);
 		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
 		Assert.isTrue(differenceGeometries.size() == 3);
 	}