+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Geometry validations
 - Asynchronous tasks
 
+## [0.2.1]SNAPSHOT - 2017-03-10
+
+### Added
+- Intersection, Difference and Symmetric Difference returns FlatGeometries.
+- improve performance
+
 ## [0.2.0]SNAPSHOT - 2017-03-09
 
 ### Added
@@ -20,8 +26,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Implemented resources for JTS operations:
 	- Difference
-	- Symetric Difference 
-- Added support for Points and LineStrings in intersection, Difference and Symetric Difference operations. Beware with tolerance to obtain best results
+	- Symmetric Difference 
+- Added support for Points and LineStrings in intersection, Difference and Symmetric Difference operations. Beware with tolerance to obtain best results
 
 ## [0.0.1]SNAPSHOT - 2017-03-06